@@ -11,11 +11,11 @@
     "javascript-color-gradient": "^2.4.4"
   },
   "devDependencies": {
-    "@types/react": "^18.0.27",
+    "@types/react": "^18.3.1",
     "process": "^0.11.10",
     "prettier": "2.5.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "typescript": "5.3.3"
   },
   "lint-staged": {
